Add __contains__ to MapKeys
I ran into this when I tried to pass `.keys()` to something declared
to take a `Collection`, and it failed because it lacked
`__contains__`. (I didn't actually care about `__contains__`, but mypy
did.)

So I came to add it to the stubs and discovered that it *was* actually
missing, and since the fallback is woefully slow I went and
implemented it.